更新版本至1.3.6
1，修复运行中存在的变量缺失导致的安装问题。
2，增大timeout的时间。

+++ CleanAll.sh
@@ -15,5 +15,18 @@ su - ${USER_NAME} -c "onclean -ky"
 ps -ef | awk '/oninit/ && $3==1{print "kill -9 "$2}' | sh
 userdel -rf ${USER_NAME} 2>/dev/null
 groupdel ${USER_NAME} 2>/dev/null
-rm -rf ${INSTALL_DIR} 2>/dev/null
-rm -rf ${WORKDIR}/install 2>/dev/null
+
+if [ -d $INSTALL_DIR -a ! x$INSTALL_DIR = x/ ]; then
+  read -p "Drop directory [$INSTALL_DIR]? [Y/N] " ISDROP 
+  if [ x"${ISDROP}" = x"Y" -o x"${ISDROP}" = x"y" ];then 
+    rm -rf ${INSTALL_DIR} 2>/dev/null
+  else
+    echo "Directory [$INSTALL_DIR] not clean."
+  fi
+fi
+
+if [ -d ${WORKDIR}/install ]; then
+  rm -rf ${WORKDIR}/install 2>/dev/null
+fi
+
+exit 0
